lambda/requirements.txt
boto3
PyNaCl
